+++ examples/compile.rs
@@ -103,7 +103,7 @@ fn main() -> ExitCode {
             let compiled_code = Compiler::compile(&src, &ast);
             generating_asm_sub_step.sub_step(&GENERATING_ASM, Some(&artifacts.asm_path));
             if let Err(err) = std::fs::write(&artifacts.asm_path, compiled_code) {
-                let error = error::Msg { kind: &COULD_NOT_WRITE_COMPILED_CODE, message: &err };
+                let error = error::MsgSimple { kind: &COULD_NOT_WRITE_COMPILED_CODE, message: &err };
                 eprintln!("{error}");
                 return ExitCode::FAILURE;
             }
@@ -130,7 +130,7 @@ fn main() -> ExitCode {
             match assembler_result {
                 Ok(output) => {
                     if !output.status.success() {
-                        let error = error::Msg {
+                        let error = error::MsgSimple {
                             kind: &ASSEMBLING_ERROR,
                             message: &String::from_utf8_lossy(&output.stderr),
                         };
@@ -143,7 +143,7 @@ fn main() -> ExitCode {
                     }
                 },
                 Err(err) => {
-                    let error = error::Msg { kind: &COULD_NOT_RUN_ASSEMBLER, message: &err };
+                    let error = error::MsgSimple { kind: &COULD_NOT_RUN_ASSEMBLER, message: &err };
                     eprintln!("{error}");
                     return ExitCode::FAILURE;
                 },
@@ -159,7 +159,7 @@ fn main() -> ExitCode {
         match linker_result {
             Ok(output) => {
                 if !output.status.success() {
-                    let error = error::Msg {
+                    let error = error::MsgSimple {
                         kind: &LINKING_ERROR,
                         message: &String::from_utf8_lossy(&output.stderr),
                     };
@@ -172,7 +172,7 @@ fn main() -> ExitCode {
                 }
             },
             Err(err) => {
-                let error = error::Msg { kind: &COULD_NOT_RUN_LINKER, message: &err };
+                let error = error::MsgSimple { kind: &COULD_NOT_RUN_LINKER, message: &err };
                 eprintln!("{error}");
                 return ExitCode::FAILURE;
             },