+++ src/main.rs
@@ -143,7 +143,7 @@ fn main() -> ExitCode {
                 verbosity,
             );
             if let Err(err) = std::fs::write(&artifacts.asm_path, compiled_code) {
-                let error = error::Msg { kind: &COULD_NOT_WRITE_COMPILED_CODE, message: &err };
+                let error = error::MsgSimple { kind: &COULD_NOT_WRITE_COMPILED_CODE, message: &err };
                 eprintln!("{error}");
                 return ExitCode::FAILURE;
             }
@@ -189,7 +189,7 @@ fn main() -> ExitCode {
             match assembler_result {
                 Ok(output) => {
                     if !output.status.success() {
-                        let error = error::Msg {
+                        let error = error::MsgSimple {
                             kind: &ASSEMBLING_ERROR,
                             message: &String::from_utf8_lossy(&output.stderr),
                         };
@@ -202,7 +202,7 @@ fn main() -> ExitCode {
                     }
                 },
                 Err(err) => {
-                    let error = error::Msg { kind: &COULD_NOT_RUN_ASSEMBLER, message: &err };
+                    let error = error::MsgSimple { kind: &COULD_NOT_RUN_ASSEMBLER, message: &err };
                     eprintln!("{error}");
                     return ExitCode::FAILURE;
                 },
@@ -218,7 +218,7 @@ fn main() -> ExitCode {
         match linker_result {
             Ok(output) => {
                 if !output.status.success() {
-                    let error = error::Msg {
+                    let error = error::MsgSimple {
                         kind: &LINKING_ERROR,
                         message: &String::from_utf8_lossy(&output.stderr),
                     };
@@ -231,7 +231,7 @@ fn main() -> ExitCode {
                 }
             },
             Err(err) => {
-                let error = error::Msg { kind: &COULD_NOT_RUN_LINKER, message: &err };
+                let error = error::MsgSimple { kind: &COULD_NOT_RUN_LINKER, message: &err };
                 eprintln!("{error}");
                 return ExitCode::FAILURE;
             },
@@ -260,7 +260,7 @@ fn main() -> ExitCode {
             }
         },
         Err(err) => {
-            let error = error::Msg { kind: &COULD_NOT_RUN_EXECUTABLE, message: &err };
+            let error = error::MsgSimple { kind: &COULD_NOT_RUN_EXECUTABLE, message: &err };
             eprintln!("{error}");
             return ExitCode::FAILURE;
         },