+++ CHANGELOG.md
@@ -56,6 +56,7 @@ but may switch to [CalVer Versioning](https://calver.org/) in the future.
     - Introduced `ArgParser`: does not require the first argument to be the executable file
     - `Command::Help` no longer contains the executable name
     - `Command::Check`, `Command::Compile` and `Command::Run` now contain `&Path`
+- `error::Msg` is now `error::MsgSimple`
 <!--  -->
 - [Syntax specification](SYNTAX.ebnf) only considers syntax and not typing